fix(tracing): panic when calling Tracer
This fixes a panic when calling `Tracer` on TracerProvider with a different name than “github.com/traefik/traefik”. This happens because `t.TracerProvider` is nil.

Instead, I replaced it with `otel.GetTracerProvider()`, which returns the default global tracer provider. If that behaviour is not desired, I would recommend using `noop.NewTracerProvider()` instead.

I created a unit test to demonstrate the problem.